Ensure `guessAuthor` returns object with `name`
The `guessAuthor` function was accidentally returning an object with both `author` and `name` properties set (to `process.env.USER || process.env.USERNAME` and the value of git's `user.name` config, respectively). This ensures that the object returned from `guessAuthor` will always be an object with only the `name` and `email` keys, that it will prefer the git config `user.name` property when it can find it, and that the default value for the "author" prompt in the `neon new` flow will be the name returned from `guessAuthor`.

This fixes a problem where the `process.env.USER || process.env.USERNAME` value was always used as the default for the "auther" prompt in `neon new`, even if there was a git config `user.name` setting available.