Unwrap v8::External pointer in C++
napi automatically wraps and unwraps the `data` pointer on functions
created through `napi_create_function`, so we don't need to use the
`napi_create_external` API in neon-runtime. Instead we can work with the
raw pointers. This aligns the C++ and n-api APIs so that the `neon`
crate can just assume associated data for a function is always a raw
pointer.