feat(cli): Add `--crate` option to specify the Rust crate
This commit adds a new option to the `build` and `clean` commands: `--crate`. This option allows to manually configure the directory containing the Rust crate.

By default, the CLI always looks into the `native` sub-directory to find the Rust crate. Using this flag, you can easily change this directory. In particular, it allows to place the `Cargo.toml` next to the `package.json` and use `neon build --crate .`.