fix(neon): Emit an unhandled rejection instead of an uncaught exception to avoid node bug
Due to a bug in the default `unhandledRejection` handler, Node crashes when triggering an uncaught exception from an error domain. Since Neon asynchronous callbacks are implicitly in an error domain, triggering an uncaught exception crashed the process. Using `unhandledRejection` is awkward, but avoids the issue.

In the future, it is possible to switch back to `unhandledRejection` without a breaking change.

1. Add a feature flag for using `unhandledRejection` instead of `unhandledRejection`
2. Update `create-neon` to use the new feature flag, making the change for new users
3. Document the change and encourage existing users to switch

Another option is to decide that this is a "bug fix" and doesn't require a breaking change even though semantics are changing. It is unlikely to impact may users, and only users on Node versions before 15. A feature flag could be added for restoring the previous behavior as an escape hatch for existing users.