neon: Relax the lifetime constraints on `TypedArray` borrows
The current lifetimes are overly restrictive. The reference is valid for as long as the handle is valid. Since the `Context` is always the
most narrow scope, it will be valid for at *least* as long as `Context`, even if the `Handle` type is dropped. This allows downcasting and
returning a borrow within the same function.