fix: More accurate lifetime bounds on scoped methods to allow valid code to compile
Prior to this change, it was impossible to escape a value in `execute_scoped` that was created in an outer scope. This is because the
function definition allowed for the outer and inner lifetimes to be entirely unrelated.

This is technically a breaking change, but in many cases the more strict bound will cause code to be more likely to compile.