Fix panic triggered by boxed Channel instances
This fixes a bug that can be triggered when there's a JsBox holding an
instance of Channel at the moment the VM stops (e.g., when quitting an
Electron app). The Channel's `drop` implementation calls `self.send`,
which panics with a SendError because we're already in the middle of
tearing everything down.