Wait for Worker in test to initialize before terminating
This fixes a sporadic panic caused by a race condition between a Worker
loading our test-specific Rust extension and the worker being shut down.
This shouldn't panic in principle, but fixing it would require
introducing a leak (see 2bf1525fe8 for the change we could have made and
decided not to).