update the cargo-cp-artifact to make windows nice (#972)
update the cargo-cp-artifact to make windows nice

Spawn has slightly different behavior on windows causing the cp command to fail when passing it additional arguments such as `-- type output.txt` (similar to `cat`). This additive change ensures that we are spawning a shell when appropriate on a windows platform. Tested cross platform.

---------

Co-authored-by: K.J. Valencik <kjvalencik@gmail.com>